Make it clear the PostgreSQL password should be supplied
If the PostgreSQL password is not explicitly supplied during chart
installation or upgrades, a random password is generated and used when
initialising the database. On future upgrades a new password is
generated in the PostgreSQL Secret resource but the database password is
not changed, so the two go out of sync. This leads to NetBox and its
rqworker failing to start.

Closes #36